refactor: move shared state to app_state module
